requirements.txt
pillow
faiss-cpu
torch
transformers
decord
boto3
numpy
